Class files removed, basic security validation added, API responses added

+++ src/main/java/com/oapps/osync/repository/IntegrationPropsRepository.java
@@ -1,11 +1,10 @@
 package com.oapps.osync.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import com.oapps.osync.entity.IntegrationPropsEntity;
 
-public interface IntegrationPropsRepository extends JpaRepository<IntegrationPropsEntity, Integer> {
+public interface IntegrationPropsRepository extends JpaRepository<IntegrationPropsEntity, Long> {
 
 	IntegrationPropsEntity findByOsyncId(Long osyncId);
